requirements.txt
flask==2.0.1
python-dotenv==0.19.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
requests==2.28.1 